Extract HTTP send file tests in a test class hierarchy
